1. Restructured package hierarchy. 2. renamed artifact id to "training".
