+++ core/src/main/java/com/vikaushi/aem/constant/Constant.java
@@ -1,4 +1,4 @@
-package com.vikaushi.aem.enablement.constant;
+package com.vikaushi.aem.constant;
 
 public class Constant {
 	public static final String SUB_SERVICE = "enablementContentEditor";
