--- core/src/main/java/com/vikaushi/aem/enablement/process/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author vikaushi
- *
- */
-package com.vikaushi.aem.enablement.process;
