+++ core/src/main/java/com/vikaushi/aem/service/ResourceResolverService.java
@@ -1,4 +1,4 @@
-package com.vikaushi.aem.enablement.service;
+package com.vikaushi.aem.service;
 
 import org.apache.sling.api.resource.LoginException;
 import org.apache.sling.api.resource.ResourceResolver;
