+++ pom.xml
@@ -6,8 +6,8 @@
     <!-- P A R E N T P R O J E C T D E S C R I P T I O N -->
     <!-- ====================================================================== -->
 
-    <groupId>com.vikaushi.aem.enablement</groupId>
-    <artifactId>enablement-project</artifactId>
+    <groupId>com.vikaushi.aem</groupId>
+    <artifactId>training</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
