+++ ui.apps/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>enablement-project.ui.apps</name>
+	<name>training.ui.apps</name>
 	<comment></comment>
 	<projects>
 	</projects>
