+++ ui.apps/.settings/org.eclipse.wst.common.component
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="enablement-project.ui.apps">
+    <wb-module deploy-name="training.ui.apps">
         <wb-resource deploy-path="/" source-path="/src/main/content/jcr_root" tag="defaultRootSource"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/test/java"/>
         <property name="context-root" value="/"/>
-        <property name="java-output-path" value="/enablement-project.ui.apps/target/classes"/>
+        <property name="java-output-path" value="/training.ui.apps/target/classes"/>
     </wb-module>
 </project-modules>
